+++ ThirdSeminar/src/main/java/sopt/org/ThirdSeminar/controller/post/PostController.java
@@ -1,11 +1,11 @@
-package sopt.org.ThirdSeminar.controller;
+package sopt.org.ThirdSeminar.controller.post;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import sopt.org.ThirdSeminar.common.dto.ApiResponseDto;
-import sopt.org.ThirdSeminar.controller.dto.request.PostRequestDto;
-import sopt.org.ThirdSeminar.controller.dto.response.PostResponseDto;
+import sopt.org.ThirdSeminar.controller.post.dto.response.ResisterResponseDto;
+import sopt.org.ThirdSeminar.controller.user.dto.request.SignupRequestDto;
 import sopt.org.ThirdSeminar.exception.SuccessStatus;
 import sopt.org.ThirdSeminar.service.PostService;
 
@@ -19,7 +19,7 @@ public class PostController {
 
     @PostMapping("/post")
     @ResponseStatus(HttpStatus.CREATED)
-    public ApiResponseDto<PostResponseDto> create(@RequestParam final Long user_id, @RequestBody @Valid final PostRequestDto request) {
+    public ApiResponseDto<ResisterResponseDto> create(@RequestParam final Long user_id, @RequestBody @Valid final SignupRequestDto request) {
         return ApiResponseDto.success(SuccessStatus.POST_REGISTER_SUCCESS, postService.create(user_id, request));
     }
 }
