+++ ThirdSeminar/src/main/java/sopt/org/ThirdSeminar/controller/user/UserController.java
@@ -1,4 +1,4 @@
-package sopt.org.ThirdSeminar.controller;
+package sopt.org.ThirdSeminar.controller.user;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import sopt.org.ThirdSeminar.common.dto.ApiResponseDto;
-import sopt.org.ThirdSeminar.controller.dto.request.UserRequestDto;
-import sopt.org.ThirdSeminar.controller.dto.response.UserResponseDto;
+import sopt.org.ThirdSeminar.controller.post.dto.request.RegisterRequestDto;
+import sopt.org.ThirdSeminar.controller.user.dto.response.SignupResponseDto;
 import sopt.org.ThirdSeminar.exception.SuccessStatus;
 import sopt.org.ThirdSeminar.service.UserService;
 
@@ -21,7 +21,7 @@ public class UserController {
 
     @PostMapping("user/signup")
     @ResponseStatus(HttpStatus.CREATED)
-    public ApiResponseDto<UserResponseDto> create(@RequestBody @Valid final UserRequestDto request){
-        return ApiResponseDto.success(SuccessStatus.SIGNUP_SUCCESS,userService.create(request));
+    public ApiResponseDto<SignupResponseDto> create(@RequestBody @Valid final RegisterRequestDto request) {
+        return ApiResponseDto.success(SuccessStatus.SIGNUP_SUCCESS, userService.create(request));
     }
 }
