+++ ThirdSeminar/src/main/java/sopt/org/ThirdSeminar/controller/user/dto/request/SignupRequestDto.java
@@ -1,4 +1,4 @@
-package sopt.org.ThirdSeminar.controller.dto.request;
+package sopt.org.ThirdSeminar.controller.user.dto.request;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -7,12 +7,11 @@ import sopt.org.ThirdSeminar.domain.User;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Pattern;
-import java.util.Optional;
 
 @Getter
 @RequiredArgsConstructor
 @Builder
-public class PostRequestDto {
+public class SignupRequestDto {
 
     @NotBlank
     @Pattern(regexp = "^[가-힣a-zA-Z]{2,30}$", message = "제목 형식에 맞지 않습니다.")
