+++ ThirdSeminar/src/main/java/sopt/org/ThirdSeminar/service/UserService.java
@@ -2,8 +2,8 @@ package sopt.org.ThirdSeminar.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import sopt.org.ThirdSeminar.controller.dto.request.UserRequestDto;
-import sopt.org.ThirdSeminar.controller.dto.response.UserResponseDto;
+import sopt.org.ThirdSeminar.controller.post.dto.request.RegisterRequestDto;
+import sopt.org.ThirdSeminar.controller.user.dto.response.SignupResponseDto;
 import sopt.org.ThirdSeminar.domain.User;
 import sopt.org.ThirdSeminar.infrastructure.UserRepository;
 
@@ -15,8 +15,8 @@ public class UserService {
     private final UserRepository userRepository;
 
     @Transactional
-    public UserResponseDto create(UserRequestDto request){
-        User user= User.builder()
+    public SignupResponseDto create(RegisterRequestDto request) {
+        User user = User.builder()
                 .email(request.getEmail())
                 .nickname(request.getNickname())
                 .password(request.getPassword())
@@ -24,6 +24,6 @@ public class UserService {
 
         userRepository.save(user);
 
-        return UserResponseDto.of(user.getId(),user.getNickname());
+        return SignupResponseDto.of(user.getId(), user.getNickname());
     }
 }
